push json() impl to interface
