+++ typedb/concept/thing/attribute.py
@@ -20,7 +20,6 @@
 #
 from abc import ABC
 from datetime import datetime
-from typing import Mapping, Union
 
 import typedb_protocol.common.concept_pb2 as concept_proto
 
@@ -40,13 +39,6 @@ class _Attribute(Attribute, _Thing, ABC):
     def as_attribute(self) -> "Attribute":
         return self
 
-    def json(self) -> Mapping[str, Union[str, int, float, bool]]:
-        return {
-            "type": self.get_type().get_label().name(),
-            "value_type": str(self.get_type().get_value_type()),
-            "value": self.get_value(),
-        }
-
 
 class _RemoteAttribute(RemoteAttribute, _RemoteThing, ABC):
 
@@ -56,13 +48,6 @@ class _RemoteAttribute(RemoteAttribute, _RemoteThing, ABC):
     def get_owners(self, owner_type: ThingType = None):
         return (concept_proto_reader.thing(t) for rp in self.stream(attribute_get_owners_req(self.get_iid(), concept_proto_builder.thing_type(owner_type)))
                 for t in rp.attribute_get_owners_res_part.things)
-
-    def json(self) -> Mapping[str, Union[str, int, float, bool]]:
-        return {
-            "type": self.get_type().get_label().name(),
-            "value_type": str(self.get_type().get_value_type()),
-            "value": self.get_value(),
-        }
 
 
 class _BooleanAttribute(BooleanAttribute, _Attribute):
@@ -237,13 +222,6 @@ class _DateTimeAttribute(DateTimeAttribute, _Attribute):
     def as_remote(self, transaction):
         return _RemoteDateTimeAttribute(transaction, self.get_iid(), self.is_inferred(), self.get_type(), self.get_value())
 
-    def json(self) -> Mapping[str, Union[str, int, float, bool]]:
-        return {
-            "type": self.get_type().get_label().name(),
-            "value_type": str(self.get_type().get_value_type()),
-            "value": self.get_value().isoformat(timespec='milliseconds')
-        }
-
 
 class _RemoteDateTimeAttribute(RemoteDateTimeAttribute, _RemoteAttribute):
 
@@ -260,10 +238,3 @@ class _RemoteDateTimeAttribute(RemoteDateTimeAttribute, _RemoteAttribute):
 
     def as_remote(self, transaction):
         return _RemoteDateTimeAttribute(transaction, self.get_iid(), self.is_inferred(), self.get_type(), self.get_value())
-
-    def json(self) -> Mapping[str, Union[str, int, float, bool]]:
-        return {
-            "type": self.get_type().get_label().name(),
-            "value_type": str(self.get_type().get_value_type()),
-            "value": self.get_value().isoformat(timespec='milliseconds')
-        }
